==> 00_spell_checkers_test.ipynb <==
# Imports

! git clone https://github.com/ai-forever/sage.git

cd sage

! pip install .
! pip install -r requirements.txt

import os
import torch
from sage.spelling_correction import T5ModelForSpellingCorruption, RuM2M100ModelForSpellingCorrection, AvailableCorrectors

m2m_418m_corrector = RuM2M100ModelForSpellingCorrection.from_pretrained(AvailableCorrectors.m2m100_418M.value)

m2m_418m_corrector.correct("биремена в 16")

batch = ["биремена в 16",
         "надежда санько",
         "Выжить в дуба",
         "бенремена в 16",
         "Звезды в дубаи",
         "битва сиьнейших",
         "stend up"]

m2m_418m_corrector.batch_correct(batch, batch_size=2)

# deeppavlov test

! pip install deeppavlov

! pip install sacremoses kenlm

import sys

from deeppavlov import build_model, configs

CONFIG_PATH = configs.spelling_correction.levenshtein_corrector_ru

model = build_model(CONFIG_PATH, download=True)

%%time
for line in batch:
    print(line, 'corrected', model([line])[0], flush=True)

## symspell + custom dicts

batch = ["биремена в 16",
         "надежда санько",
         "Выжить в дуба",
         "бенремена в 16",
         "Звезды в дубаи",
         "битва сиьнейших",
         "stend up"]

from spelling_checker.sym_spell_servicer import SymSpellRouterServicer

spell_checker = SymSpellRouterServicer()

%%time
for line in batch:
    print(line, '\t corrected:', spell_checker.predict_single_correction(
    line,
    use_preprocessing=True,
    use_keyboard_inverter=True,
    use_correction=True
)[1], flush=True)
