==> 00_symspell_vocabulary_generation.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import sys

sys.path.insert(0, '/content/drive/MyDrive/hack_2023_11')

from collections import defaultdict
import matplotlib.pyplot as plt
import numpy as np
from nltk import ngrams
import os
import pandas as pd
import re
from tqdm import tqdm
tqdm.pandas()

from spelling_checker.sym_spell_servicer import SymSpellRouterServicer

spell_checker = SymSpellRouterServicer()

WORD_PATTERN = re.compile(r"[a-zа-яйё!&'._|\-+:]+")
WORD_STRIP = "&'_|-:"
VALID_CHARS = "!&'._|-+:"

# Load data

video_meta = pd.read_parquet("/content/drive/MyDrive/hack_2023_11/data/videos.parquet")

queries_1 = pd.read_parquet("/content/drive/MyDrive/hack_2023_11/data/automarkup.parquet")
queries_2 = pd.read_csv("/content/drive/MyDrive/hack_2023_11/data/manualmarkup.csv")

video_meta.sample(2)

queries_1.sample(2)

queries_2.sample(2)

# Aggregate video attributes

video_meta = pd.DataFrame(video_meta[["video_title", "channel_title"]].unstack().rename("text")).reset_index(drop=True)

video_meta = video_meta[video_meta["text"].progress_apply(lambda x: WORD_PATTERN.search(x) is not None)] \
    .reset_index(drop=True)

video_meta["clean_text"] = video_meta["text"] \
    .progress_apply(
    lambda x: spell_checker.predict_single_correction(
        x, use_preprocessing=True, use_keyboard_inverter=False, use_correction=False
    ))

unigrams = defaultdict(int)
for idx, sublist in tqdm(enumerate(video_meta["clean_text"])):
    for word in sublist.split():
        if WORD_PATTERN.search(word) is not None:
            if word in VALID_CHARS:
                pass
            else:
                word = word.strip(WORD_STRIP)
            unigrams[word] = unigrams.get(word, 0) + 1

bigrams = defaultdict(int)
for sublist in tqdm(video_meta["clean_text"]):
    tokens = []
    for word in sublist.split():
        if WORD_PATTERN.search(word) is not None:
            if word in VALID_CHARS:
                tokens.append(word)
            else:
                tokens.append(word.strip(WORD_STRIP))
    for bigram in ngrams(tokens, 2):
        b = " ".join(bigram)
        bigrams[b] = bigrams.get(b, 0) + 1

min_unigram = int(spell_checker.config["SYMSPELL_PARAMS"]["MIN_UNIGRAM_COUNT"])
min_bigram = int(spell_checker.config["SYMSPELL_PARAMS"]["MIN_BIGRAM_COUNT"])

unigrams = {k: v if v >= min_unigram else min_unigram for k, v in unigrams.items()}
bigrams = {k: v if v >= min_bigram else min_bigram for k, v in bigrams.items()}

video_meta.to_parquet("data/clean_videos.parquet", index=False)

# Aggregate queries

queries = pd.DataFrame(pd.concat([queries_1["query"], queries_2["query"]]).rename("text"))

del queries_1, queries_2

queries = queries[queries["text"].progress_apply(lambda x: WORD_PATTERN.search(x) is not None)] \
    .reset_index(drop=True)

queries["clean_text"] = queries["text"] \
    .progress_apply(
    lambda x: spell_checker.predict_single_correction(
        x, use_preprocessing=True, use_keyboard_inverter=False, use_correction=False
    ))

q_unigrams = defaultdict(int)
for idx, sublist in tqdm(enumerate(queries["clean_text"])):
    for word in sublist.split():
        if WORD_PATTERN.search(word) is not None:
            if word in VALID_CHARS:
                pass
            else:
                word = word.strip(WORD_STRIP)
            q_unigrams[word] = q_unigrams.get(word, 0) + 1

q_bigrams = defaultdict(int)
for sublist in tqdm(queries["clean_text"]):
    tokens = []
    for word in sublist.split():
        if WORD_PATTERN.search(word) is not None:
            if word in VALID_CHARS:
                tokens.append(word)
            else:
                tokens.append(word.strip(WORD_STRIP))
    for bigram in ngrams(tokens, 2):
        b = " ".join(bigram)
        q_bigrams[b] = q_bigrams.get(b, 0) + 1

np.quantile(list(q_unigrams.values()), .75), np.quantile(list(q_unigrams.values()), .95), np.quantile(list(q_unigrams.values()), .99)

np.quantile(list(q_bigrams.values()), .75), np.quantile(list(q_bigrams.values()), .95), np.quantile(list(q_bigrams.values()), .99)

print([k for k, v in q_unigrams.items() if v > 321])

q_unigrams = {k: v if v >= min_unigram else 321 for k, v in q_unigrams.items()}
q_bigrams = {k: v if v >= min_bigram else 62 for k, v in q_bigrams.items()}

# Merge and collect final dicts

for k, v in q_unigrams.items():
    unigrams[k] = unigrams.get(k, 0) + v
del q_unigrams

for k, v in q_bigrams.items():
    bigrams[k] = bigrams.get(k, 0) + v
del q_bigrams

len(unigrams), len(bigrams)

with open("spelling_checker/mount_files/symspell/ru.txt", "r", encoding="utf-8") as f:
    rus = f.readlines()

with open("spelling_checker/mount_files/symspell/en.txt", "r", encoding="utf-8") as f:
    eng = f.readlines()

rus = dict([i.strip().split() for i in rus])
eng = dict([i.strip().split() for i in eng])

vocab = defaultdict(int)

for k, v in unigrams.items():
    vocab[k] = int(rus.get(k, 0)) + int(eng.get(k, 0)) + v

filename = os.path.join("/content/drive/MyDrive/hack_2023_11/spelling_checker/mount_files/symspell", "unigrams.txt")

with open(filename, "w", encoding="utf-8") as f:
    for k, v in sorted(vocab.items(), key=lambda item: item[1], reverse=True):
        f.write(f"{k}${v}\n")

filename = os.path.join("/content/drive/MyDrive/hack_2023_11/spelling_checker/mount_files/symspell", "bigrams.txt")

with open(filename, "w", encoding="utf-8") as f:
    for k, v in sorted(bigrams.items(), key=lambda item: item[1], reverse=True):
        f.write(f"{k}${v}\n")
