==> 02_build_es_indexes.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

!pip install elasticsearch eng_to_ipa

! pip install faiss-gpu==1.7.2 sentence-transformers==2.2.2 numpy==1.25

! pip install numpy==1.25

! pip install fasttext

path = '../content/drive/MyDrive/JETFORK_11_2023/data/'

import faiss
import fasttext

from elasticsearch import Elasticsearch, helpers
import pandas as pd
import re
from tqdm import tqdm
import sys

sys.path.insert(0, path)


es = Elasticsearch("http://5.35.83.33:9200/", basic_auth=("elastic", ""))

ENG_WORD_PATTERN = re.compile(r"[a-z]+")

model = fasttext.load_model(path + 'model/00_fasttext_queries.bin')

video_meta = pd.read_parquet(path + "clean_videos_done.parquet")[['video_id', 'video_title', 'channel_title',
       'v_channel_type', 'v_category', 'clean_title',
       'clean_channel_title']]

video_meta = video_meta.dropna(subset=['clean_title', 'clean_channel_title'])
video_meta.columns = ['id', 'source_video_title', 'source_channel_title', 'v_channel_type',
       'v_category', 'processed_video_title', 'processed_channel_title']
video_meta = video_meta.reset_index(drop=True)

import gc
gc.collect()

from tqdm import tqdm

tqdm.pandas()

video_meta = video_meta.head(15000000)

video_meta['video-vector'] = video_meta['processed_video_title'].progress_apply(lambda x: list(model.get_sentence_vector(x)))

# Video vectors

INDEX_NAME = "videos_index__v1"
es.indices.create(index=INDEX_NAME,
                  body={
                      "mappings": {
                        "properties": {
                          "video-vector": {
                            "type": "dense_vector",
                            "dims": 70,
                            "similarity": "cosine"
                          },
                          "id": {"type": "text"},
                          "source_video_title": {"type": "text"},
                          "source_channel_title": {"type": "text"},
                          "v_channel_type": {"type": "text"},
                          "v_category": {"type": "text"},
                          "processed_video_title": {"type": "text"},
                          "processed_channel_title": {"type": "text"},
                          }
                      }
                    })

! pip install es_pandas

from es_pandas import es_pandas

ep = es_pandas("http://5.35.83.33:9200/")

ep.to_es(video_meta, INDEX_NAME, thread_count=8, chunk_size=10000)

es.indices.refresh(index=INDEX_NAME)
