==> 05_backfill.ipynb <==
from google.colab import drive

drive.mount('/content/drive')

import pandas as pd
import pickle

automarkup = pd.read_parquet(
    '/content/drive/MyDrive/hack_2023_11/data/automarkup.parquet',
    engine='fastparquet'
)

automarkup['cnt'] = automarkup.groupby(['video_id'])['video_id'].transform('size')
automarkup = automarkup.sort_values('cnt', ascending=False) \
  [['video_id']].drop_duplicates()

videos = pd.read_parquet('/content/drive/MyDrive/hack_2023_11/data/clean_videos_done.parquet')
videos.sample(2)

automarkup = automarkup.merge(
    videos[['video_id', 'video_title', 'channel_title', 'v_category', 'v_channel_type']].drop_duplicates(subset=['video_id']),
    on='video_id', how='left'
)

with open('/content/drive/MyDrive/hack_2023_11/models/backfill.pickle', 'wb') as handle:
    pickle.dump(automarkup.to_dict('records'), handle, protocol=pickle.HIGHEST_PROTOCOL)
